--- apps/web/src/entities/log/index.ts
@@ -1,5 +1,0 @@
-export {
-  PLACE_TAG_LABELS,
-  type PlaceTagValue,
-  TAG_CATEGORIES,
-} from './model/review-tags';
