+++ apps/web/src/entities/place/index.ts
@@ -1,3 +1,11 @@
 export type { PlaceCategory, PlaceCategoryValue } from './model/place-category';
 export { PLACE_CATEGORIES } from './model/place-category';
+export {
+  AtmosphereAndFocus,
+  EnvironmentAndComfort,
+  OtherTags,
+  type PlaceTagValue,
+  SeatingAndSpace,
+  WorkConvenience,
+} from './model/place-tag';
 export type { Place, PlaceLayer } from './model/types';
